requirements.txt
opencv-python>=4.7.0
mediapipe>=0.9.0
pyautogui>=0.9.53
numpy>=1.21.0
PyQt6>=6.4.0
pillow>=9.0.0
pyinstaller>=5.6.0